Refactor: 구글맵 옵션과 거리 계산 함수 분리 작업 (#72)
- google map style 옵션, getDistance 함수 utils 폴더로 분리
