+++ client/src/components/common/Map.tsx
@@ -1,6 +1,8 @@
 import { GoogleMap, LoadScript, Marker, Polyline } from '@react-google-maps/api'
 import { useState, useEffect } from 'react'
 import styles from './Map.module.scss'
+import getDistance from '../../utils/calculate'
+import { googleMapOptions, polylineOptions } from '../../utils/options'
 
 // 좌표 테스트 ing
 
@@ -18,46 +20,29 @@ function Map() {
   const [locations, setLocations] = useState<LatLng[]>([])
   const [isLoadMap, setIsLoadMap] = useState(false)
   const [distance, setDistance] = useState(0)
-  console.log('currentLocation: ', currentLocation)
 
   useEffect(() => {
-    setIsLoadMap(true)
     setApiKey(import.meta.env.VITE_GOOGLE_MAPS_API_KEY || '')
-    // navigator.geolocation.getCurrentPosition(
-    //   position => {
-    //     if (position) {
-    //       const location = {
-    //         lat: position.coords.latitude,
-    //         lng: position.coords.longitude,
-    //       }
-    //       setCurrentLocation(location)
-    //     }
-    //   },
-    //   function (error) {
-    //     alert(`Error occurred. Error code: ${error.code}`)
-    //   },
-    //   { timeout: 5000 }
-    // )
+    navigator.geolocation.getCurrentPosition(
+      position => {
+        if (position) {
+          const location = {
+            lat: position.coords.latitude,
+            lng: position.coords.longitude,
+          }
+          setCurrentLocation(location)
+        }
+      },
+      function (error) {
+        alert(`Error occurred. Error code: ${error.code}`)
+      },
+      { timeout: 5000 }
+    )
+    setIsLoadMap(true)
     return () => {
       setIsLoadMap(false)
     }
   }, [])
-
-  function getDistance(lat1: number, lng1: number, lat2: number, lng2: number): number {
-    console.log('getDistance 실행')
-    function deg2rad(deg: number): number {
-      return deg * (Math.PI / 180)
-    }
-    const R = 6371 // Radius of the earth in km
-    const dLat = deg2rad(lat2 - lat1) // deg2rad below
-    const dLon = deg2rad(lng2 - lng1)
-    const a =
-      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-      Math.cos(deg2rad(lat1)) * Math.cos(deg2rad(lat2)) * Math.sin(dLon / 2) * Math.sin(dLon / 2)
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a))
-    const d = R * c * 1000 // Distance in m
-    return d
-  }
 
   function watchLocation() {
     let prevLocation: LatLng | undefined // 이전 위치
@@ -72,19 +57,16 @@ function Map() {
             lat: position.coords.latitude,
             lng: position.coords.longitude,
           }
-          console.log('location: ', location)
           // 현재 위치 갱신
           setCurrentLocation(() => location) // state
           // 직전 좌표가 없으면 현재 watchPosition으로 받은 location을 locations 배열에 추가
           if (!prevLocation) {
             setLocations([location])
             prevLocation = location
-            console.log('prevLocation: ', prevLocation)
           }
           // 직전 좌표가 있으면 locations 배열에 현재 위치 추가
-          console.log('location: ', location, 'prevLocation: ', prevLocation)
+          // currentLocation에서 location으로 변경
           if (location && prevLocation) {
-            console.log('if문 안으로 들어옴')
             // 직전 좌표가 존재하면 현재 좌표와 직전 좌표의 거리를 계산
             const distanceFromLastPosition = getDistance(
               location.lat,
@@ -93,7 +75,6 @@ function Map() {
               prevLocation.lng
             )
 
-            console.log('distanceFromLastPosition: ', distanceFromLastPosition)
             // distance 상태를 갱신
             setDistance(distanceFromLastPosition)
             if (distanceFromLastPosition >= 5) {
@@ -125,62 +106,10 @@ function Map() {
             mapContainerStyle={containerStyle}
             zoom={14}
             center={currentLocation}
-            options={{
-              styles: [
-                {
-                  // 색상
-                  featureType: 'all',
-                  stylers: [
-                    {
-                      saturation: -100,
-                    },
-                  ],
-                },
-                {
-                  // 물 색상
-                  featureType: 'water',
-                  stylers: [
-                    {
-                      color: '#7dcdcd',
-                    },
-                  ],
-                },
-                {
-                  // 건물 이름 가리기
-                  featureType: 'all',
-                  elementType: 'labels',
-                  stylers: [{ visibility: 'off' }],
-                },
-
-                {
-                  // 지도 단순화하기
-                  featureType: 'road',
-                  elementType: 'geometry',
-                  stylers: [
-                    { visibility: 'simplified' },
-                    { hue: '#000000' },
-                    { saturation: -50 },
-                    { lightness: -15 },
-                    { weight: 1.5 },
-                  ],
-                },
-              ],
-              mapTypeControl: false, // 지도 위성 끄기
-              streetViewControl: false, // 사람 모양 끄기
-              gestureHandling: 'greedy', // 한 손 가락으로 지도 핸들링
-            }}
+            options={googleMapOptions}
           >
             {currentLocation && <Marker position={currentLocation} />}
-            {locations.length > 1 && (
-              <Polyline
-                path={locations}
-                options={{
-                  strokeColor: '#55de50',
-                  strokeOpacity: 1,
-                  strokeWeight: 6,
-                }}
-              />
-            )}
+            {locations.length > 1 && <Polyline path={locations} options={polylineOptions} />}
           </GoogleMap>
         </LoadScript>
       ) : null}
